+++ src/components/ThumbnailList/__tests__/ThumbnailList.test.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import PhotoList from '..';
+import ThumbnailList from '..';
 import PLACE from '../../../mockData/PLACE';
 import utils from '../../../utils';
 
-describe('PhotoList', () => {
+describe('ThumbnailList', () => {
   const photos = PLACE.photoReferences.map((photoReference) => {
     return {
       uri: utils.app.getGooglePlacesPhotoURI(photoReference),
@@ -14,7 +14,7 @@ describe('PhotoList', () => {
 
   describe('renders', () => {
     it('renders with minimum required props', () => {
-      const component = renderer.create(<PhotoList data={photos} />);
+      const component = renderer.create(<ThumbnailList data={photos} />);
 
       expect(component).toMatchSnapshot();
     });